Implement basic diagram drawing

+++ core/src/networkmodeling/core/IpAddress.java
@@ -3,7 +3,7 @@ package networkmodeling.core;
 import java.io.Serializable;
 
 public class IpAddress implements Serializable {
-    public IpAddress(byte octet1, byte octet2, byte octet3, byte octet4) {
+    public IpAddress(short octet1, short octet2, short octet3, short octet4) {
         this.octet1 = octet1;
         this.octet2 = octet2;
         this.octet3 = octet3;
@@ -34,8 +34,8 @@ public class IpAddress implements Serializable {
     }
     
     
-    private final byte octet1;
-    private final byte octet2;
-    private final byte octet3;
-    private final byte octet4;
+    private final short octet1;
+    private final short octet2;
+    private final short octet3;
+    private final short octet4;
 }
